Fixed after review of Labs 1-5.

+++ src/Lab1/Main.java
@@ -26,8 +26,8 @@ public class Main {
         System.out.println(resultSecondExpression);
 
         //task 3
-        int number = 10500;
-        int result = (number / 10) / 10;
+        double number = 10500;
+        double result = (number / 10) / 10;
 
         System.out.println(result);
