+++ src/Lab2/Calculator.java
@@ -42,7 +42,7 @@ public class Calculator {
 
     // "/"
     public double division(int a, int b) {
-        return a / b;
+        return (double) a / b;
     }
 
     public double division(double a, double b) {
@@ -50,6 +50,6 @@ public class Calculator {
     }
 
     public double division(long a, long b) {
-        return a / b;
+        return (double) a / b;
     }
 }
